+++ src/components/Message/MessageCommerce.tsx
@@ -122,22 +122,9 @@ const MessageCommerceWithContext = <
       <div className='str-chat__message-commerce-inner'>
         <>
           {<MessageOptions displayLeft={false} displayReplies={false} theme='commerce' />}
-          {hasReactions && !showDetailedReactions && isReactionEnabled && (
-            <ReactionsList
-              own_reactions={message.own_reactions}
-              reaction_counts={message.reaction_counts || undefined}
-              reactions={message.latest_reactions}
-            />
-          )}
+          {hasReactions && !showDetailedReactions && isReactionEnabled && <ReactionsList />}
           {showDetailedReactions && isReactionEnabled && (
-            <ReactionSelector
-              detailedView
-              latest_reactions={message.latest_reactions}
-              own_reactions={message.own_reactions}
-              reaction_counts={message.reaction_counts || undefined}
-              ref={reactionSelectorRef}
-              reverse={false}
-            />
+            <ReactionSelector ref={reactionSelectorRef} />
           )}
         </>
         {message.attachments && (